Phase 1.4+1.5: PHYS_TO_VIRT for VGA framebuffer, uintptr_t buddy, arch-pointer-size user stack
- vga.c / video.c use PHYS_TO_VIRT(0xB8000) instead of hardcoded 0xC00B8000
- buddy.c drops its local 0xC0000000 KERNEL_VMA copy + 32-bit phys helpers,
  uses arch/paging.h macros and uintptr_t base_phys
- setup_initial_stack builds the argc/argv/envp pointer block with
  sizeof(uintptr_t) slots (i386 unchanged), do_exec uses uintptr_t esp
i386 regression boot verified.

+++ include/kernel/process.h
@@ -44,9 +44,9 @@ int copy_exec_strings(char **user_ptrs, exec_strings_t *out);
  * @param esp          Out: initial user ESP pointing at argc
  * @return 0 on success, -1 if the block does not fit
  */
-int setup_initial_stack(void *stack_pages, uint32_t stack_vaddr,
+int setup_initial_stack(void *stack_pages, uintptr_t stack_vaddr,
                         exec_strings_t *argv, exec_strings_t *envp,
-                        uint32_t *esp);
+                        uintptr_t *esp);
 
 /* =========================================================================
  * Process Management System Calls
